changes to apis as discussed

+++ backend/tables/views.py
@@ -101,7 +101,7 @@ def instruments_list(request):
         nextPage = 1
         previousPage = 1
         instruments = Instrument.objects.all()
-        return get_page_response(instruments, request, InstrumentReadSerializer, "instruments", nextPage, previousPage)
+        return get_page_response(instruments, request, ListInstrumentReadSerializer, "instruments", nextPage, previousPage)
 
     elif request.method == 'POST':
         # get model pk from vendor and model number
@@ -111,7 +111,7 @@ def instruments_list(request):
             model = ItemModel.objects.get(vendor=vendor, model_number=model_number)
             request.data['item_model'] = model.pk
         except ItemModel.DoesNotExist:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
+            return Response("Vendor/model number does not exist.", status=status.HTTP_400_BAD_REQUEST)
         # add new instrument using itemmodel
         serializer = InstrumentWriteSerializer(data=request.data)
         if serializer.is_valid():
@@ -133,13 +133,21 @@ def instruments_detail(request, pk):
         return Response(status=status.HTTP_404_NOT_FOUND)
 
     if request.method == 'GET':
-        serializer = InstrumentReadSerializer(instrument, context={'request': request})
+        serializer = DetailInstrumentReadSerializer(instrument, context={'request': request})
         return Response(serializer.data)
 
     elif request.method == 'PUT':
-        # fill in immutable fields
-        request.data['item_model'] = instrument.item_model.pk
-        request.data['serial_number'] = instrument.serial_number
+        if 'vendor' in request.data or 'model_number' in request.data:
+            vendor = request.data['vendor'] if 'vendor' in request.data else instrument.item_model.vendor
+            model_number = request.data['model_number'] if 'model_number' in request.data else instrument.item_model.model_number
+            try:
+                item_model = ItemModel.objects.get(vendor=vendor, model_number=model_number)
+                request.data['item_model'] = item_model.pk
+            except ItemModel.DoesNotExist:
+                return Response("Vendor/model number does not exist.", status=status.HTTP_400_BAD_REQUEST)
+        else:
+            request.data['item_model'] = instrument.item_model.pk
+        if 'serial_number' not in request.data: request.data['serial_number'] = instrument.serial_number
         serializer = InstrumentWriteSerializer(instrument, data=request.data, context={'request': request})
         if serializer.is_valid():
             serializer.save()
@@ -185,13 +193,12 @@ def models_detail(request, pk):
         return Response(status=status.HTTP_404_NOT_FOUND)
 
     if request.method == 'GET':
-        serializer = ItemModelSerializer(model, context={'request': request})
+        serializer = DetailItemModelSerializer(model, context={'request': request})
         return Response(serializer.data)
 
     elif request.method == 'PUT':
-        # fill in immutable fields
-        request.data['vendor'] = model.vendor
-        request.data['model_number'] = model.model_number
+        if 'vendor' not in request.data: request.data['vendor'] = model.vendor
+        if 'model_number' not in request.data: request.data['model_number'] = model.model_number
         serializer = ItemModelSerializer(model, data=request.data, context={'request': request})
         if serializer.is_valid():
             serializer.save()
@@ -199,8 +206,11 @@ def models_detail(request, pk):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
-        model.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if len(model.instrument_set.all()) > 0:
+            return Response("Cannot delete model with instrument instances.", status=status.HTTP_400_BAD_REQUEST)
+        else:
+            model.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 # USERS
